[TASK] Split up Utility class in several classes
This change splits up the Utility class
so the Converter can be used without having
any filesystem access in a different environment.
