+++ Tests/Fixtures/sample_extension/composer.json
@@ -1,19 +1,20 @@
 {
-  "name": "b13/test-package",
-  "type": "typo3-cms-extension",
-  "description": "Sample composer.json for running tests",
-  "license": "GPL-2.0-or-later",
-  "authors": [
-    {
-      "name": "Sample name",
-      "role": "Developer",
-      "homepage": "https://b13.com"
+    "name": "b13/test-package",
+    "type": "typo3-cms-extension",
+    "description": "Sample composer.json for running tests",
+    "license": "GPL-2.0-or-later",
+    "authors": [
+        {
+            "name": "Sample name",
+            "role": "Developer",
+            "homepage": "https://b13.com"
+        }
+    ],
+    "require": [],
+    "extra": {
+        "branch-alias": {
+            "dev-master": "11.2.x-dev"
+        },
+        "typo3/cms": []
     }
-  ],
-  "require": {},
-  "extra": {
-    "branch-alias": {
-      "dev-master": "11.2.x-dev"
-    }
-  }
 }
